+++ includes/class-tmsm-aquatonic-course-booking-activator.php
@@ -49,7 +49,6 @@ class Tmsm_Aquatonic_Course_Booking_Activator {
 
 		$charset_collate = $wpdb->get_charset_collate();
 
-		// sql to create your table
 		// NOTICE that:
 		// 1. each field MUST be in separate line
 		// 2. There must be two spaces between PRIMARY KEY and its name
@@ -58,6 +57,7 @@ class Tmsm_Aquatonic_Course_Booking_Activator {
 		$sql = "CREATE TABLE " . $table_name . " (
         booking_id INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,
         token VARCHAR(50) NOT NULL DEFAULT '',
+        barcode VARCHAR(20) NOT NULL DEFAULT '',
         firstname VARCHAR(50) NOT NULL,
         lastname VARCHAR(50) NOT NULL,
         email VARCHAR(100) NOT NULL,