+++ tmsm-aquatonic-course-booking.php
@@ -16,7 +16,7 @@
  * Plugin Name:       TMSM Aquatonic Course Booking
  * Plugin URI:        https://www.github.com/thermesmarins/tmsm-aquatonic-course-booking/
  * Description:       Aquatonic Booking
- * Version:           1.1.0
+ * Version:           1.1.1
  * Author:            Nicolas Mollet
  * Author URI:        https://github.com/nicomollet
  * License:           GPL-3.0
@@ -34,11 +34,9 @@ if ( ! defined( 'WPINC' ) ) {
 
 /**
  * Currently plugin version.
- * Start at version 1.0.0 and use SemVer - https://semver.org
- * Rename this for your plugin and update it as you release new versions.
  */
-define( 'TMSM_AQUATONIC_COURSE_BOOKING_VERSION', '1.1.0' );
-define( 'TMSM_AQUATONIC_COURSE_BOOKING_DB_VERSION', '1' );
+define( 'TMSM_AQUATONIC_COURSE_BOOKING_VERSION', '1.1.1' );
+define( 'TMSM_AQUATONIC_COURSE_BOOKING_DB_VERSION', '2' );
 
 /**
  * The code that runs during plugin activation.