+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add barcode field to course sql table
+
 ### 1.1.0: March 9th, 2021
 * Dialog Insight API refactoring
 * Dialog Insight: mark booking as arrived